update type of getNum

+++ src/main/java/org/gsc/core/wrapper/AssetIssueWrapper.java
@@ -85,7 +85,7 @@ public class AssetIssueWrapper implements ProtoWrapper<AssetIssueContract> {
     return name + "_" + order;
   }
 
-  public int getNum() {
+  public long getNum() {
     return this.assetIssueContract.getNum();
   }
